fixing diff
