+++ types/zchat-browser/index.d.ts
@@ -1,12 +1,12 @@
-// Type definitions for non-npm package zendesk-chat-browser 1.81
+// Type definitions for non-npm package zChat 1.81
 // Project: https://api.zopim.com/web-sdk/#introduction
 // Definitions by: David Copley <https://github.com/davidcopley>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.1
 
-export as namespace zChat;
+export as namespace zchat;
 
-interface InitProps {
+export interface InitProps {
     account_key: string;
     authentication?: {
         jwt_fn?: (callback: (jwt: string) => void) => void;
@@ -59,16 +59,14 @@ export function sendFile(file: File, callback?: (err: SendFileErrorMessage, data
 }) => void): void;
 
 export function sendOfflineMsg(
-    
     options: {
-            name: string;
+        name: string;
         email: string;
         phone?: string;
         message: string;
         department?: number;
     },
     callback: (err: Error) => void
-
 ): void;
 
 export function addTags(tags: string[], callback?: (err: Error) => void): void;
@@ -100,35 +98,35 @@ export namespace ChatEvent {
 
     type ChatEventData =
         | BaseChatEventData & {
-              type: 'chat.msg';
-              msg: string;
-              options: string[];
-              structured_msg: StructuredMessage;
-          }
+        type: 'chat.msg';
+        msg: string;
+        options: string[];
+        structured_msg: StructuredMessage;
+    }
         | BaseChatEventData & {
-              type: 'chat.file';
-              attachment: Attachment;
-              deleted: boolean;
-          }
+        type: 'chat.file';
+        attachment: Attachment;
+        deleted: boolean;
+    }
         | BaseChatEventData & {
-              type: 'chat.memberjoin';
-          }
+        type: 'chat.memberjoin';
+    }
         | BaseChatEventData & {
-              type: 'chat.memberleave';
-          }
+        type: 'chat.memberleave';
+    }
         | BaseChatEventData & {
-              type: 'chat.request.rating';
-          }
+        type: 'chat.request.rating';
+    }
         | BaseChatEventData & {
-              type: 'chat.rating';
-              rating?: string;
-              new_rating?: string;
-          }
+        type: 'chat.rating';
+        rating?: string;
+        new_rating?: string;
+    }
         | BaseChatEventData & {
-              type: 'chat.comment';
-              comment?: string;
-              new_comment?: string;
-          };
+        type: 'chat.comment';
+        comment?: string;
+        new_comment?: string;
+    };
 
     interface Action {
         type: 'QUICK_REPLY_ACTION' | 'LINK_ACTION';
@@ -162,24 +160,24 @@ export namespace ChatEvent {
 
     type StructuredMessage =
         | {
-              type: 'QUICK_REPLIES';
-              msg: string;
-              quick_replies: Button[];
-          }
+        type: 'QUICK_REPLIES';
+        msg: string;
+        quick_replies: Button[];
+    }
         | {
-              type: 'PANEL_TEMPLATE';
-              panel: Panel;
-              buttons: Button[];
-          }
+        type: 'PANEL_TEMPLATE';
+        panel: Panel;
+        buttons: Button[];
+    }
         | {
-              type: 'PANEL_TEMPLATE_CAROUSEL';
-              items: PanelTemplate[];
-          }
+        type: 'PANEL_TEMPLATE_CAROUSEL';
+        items: PanelTemplate[];
+    }
         | {
-              type: 'LIST_TEMPLATE';
-              items: ListItem[];
-              buttons: Button[];
-          };
+        type: 'LIST_TEMPLATE';
+        items: ListItem[];
+        buttons: Button[];
+    };
 }
 
 export type SendFileErrorMessage =
@@ -230,20 +228,20 @@ export type EventName =
 export type EventData =
     | ChatEvent.ChatEventData
     | {
-          type: 'chat.queue_position';
-          nick: string;
-          queue_position: number;
-      }
+    type: 'chat.queue_position';
+    nick: string;
+    queue_position: number;
+}
     | {
-          type: 'typing';
-          nick: string;
-          typing: boolean;
-      }
+    type: 'typing';
+    nick: string;
+    typing: boolean;
+}
     | {
-          type: 'last_read';
-          nick: string;
-          timestamp: number;
-      }
+    type: 'last_read';
+    nick: string;
+    timestamp: number;
+}
     | 'online'
     | 'away'
     | 'offline'
